requirements.txt
dash~=1.20.0
dash_bootstrap_components
flask_caching
plotly~=4.14.3
numpy~=1.20.1
pandas~=1.2.3
boto3~=1.17.35
pyyaml~=5.4.1
scikit-image~=0.18.1
tifffile~=2021.3.17
imagecodecs
matplotlib
jmespath~=0.10.0
Pillow~=8.1.2

yaml~=0.2.5
paramiko~=2.7.2
scp~=0.13.3
botocore~=1.20.35
python-dateutil~=2.8.1
